Refactor database module
Moving database module to its own package to keep the project structure clean.
